==> Bulk/Human/4_Enrichment_GSEA_results_PTEN_protein_loss_vs_presence_Hallmark_BP_CC_REACTOME.R <==
################################################################################
#| GSEA RESULTS VISUALIZATION - PTEN PROTEIN LOSS VS INTACT ###########
################################################################################
#| Date: 12/12/2024
#| Author: Ivana Rondon Lorefice
#|
#| Description:
#| This script summarizes and visualizes results from Gene Set Enrichment 
#| Analysis (GSEA) across multiple databases when comparing PTEN protein loss 
#| vs intact in prostate cancer patients. Analyses are performed in Basurto’s 
#| cohort (147 PCa samples).
#|
#| Workflow:
#|   1) Import GSEA outputs from multiple collections:
#|        - HALLMARK gene sets
#|        - Gene Ontology: Biological Process (GO:BP)
#|        - Gene Ontology: Cellular Component (GO:CC)
#|        - Gene Ontology: Molecular Function (GO:MF)
#|        - REACTOME pathways
#|   2) Clean and harmonize pathway names (remove prefixes, underscores).
#|   3) Export processed tables to Excel for record-keeping.
#|   4) Visualize top-enriched terms using:
#|        - Combined dotplots across datasets (NES values vs dataset/term).
#|        - Individual dataset dotplots (NES vs pathway, sized by NES, filled by p-value).
#|        - Bubble plots highlighting leading terms per collection.
#|   5) Repeat for additional phenotype contrasts (e.g., "Responders").
#|
#| Inputs:
#|   - GSEA result tables (TSV files) generated by GSEA software for each 
#|     database (Hallmark, GO:BP/CC/MF, Reactome).
#|
#| Outputs:
#|   - Cleaned and exported enrichment tables (Excel format).
#|   - Multi-dataset dotplot summaries of top NES pathways.
#|   - Individual dotplots and bubble plots per database (Hallmark, Reactome, GO:BP/CC/MF).
#|   - Figures saved as PDF for inclusion in reports or publications.
#|
#| Notes:
#| - NES = Normalized Enrichment Score from GSEA.
#| - NOM.p.val is used to scale dot colors.
#| - Plots use ggplot2 + viridis or Spectral palettes.
################################################################################

################################################################################
#|  LIBRARIES AND DATA
################################################################################
suppressMessages(library(ggplot2))
suppressMessages(library(viridis))
suppressMessages(library(WGCNA))
suppressMessages(library(gprofiler2))
suppressMessages(library(ggplot2))
suppressMessages(library(DESeq2))
suppressMessages(library(clusterProfiler))
suppressMessages(library(enrichplot))
suppressMessages(library(corrplot))
suppressMessages(library(ggpubr))

#| Setwd
setwd("X:/irondon/Project_AC-45_RNAseq-FFPE/RNAseq/4_STEP_DATA_ANALYSIS/4_Enrichment_Analysis/GSEA PTEN loss vs intact Basurto/Results/")

#| For plots
theme_set(theme_classic()) 

#| Loading results from GSEA
GSEA_HALLMARK <- read.table("my_analysis_Hallmark.Gsea.1681820629869/gsea_report_for_loss_1681820629869.tsv", sep = "\t", header = T)
GSEA_BP <- read.table("my_analysis_BP.Gsea.1681732378658/gsea_report_for_loss_1681732378658.tsv", sep = "\t", header = T)
GSEA_CC <- read.table("my_analysis_CC.Gsea.1681732346694/gsea_report_for_loss_1681732346694.tsv", sep = "\t", header = T)
GSEA_MF <- read.table("my_analysis_MF.Gsea.1681732409987/gsea_report_for_loss_1681732409987.tsv", sep = "\t", header = T)
GSEA_REACTOME <- read.table("my_analysis_REACTOME.Gsea.1681813676558/gsea_report_for_loss_1681813676558.tsv", sep = "\t",  header = T)

GSEA_HALLMARK$NAME <- gsub("HALLMARK_","",GSEA_HALLMARK$NAME)
GSEA_BP$NAME <- gsub("GOBP_","",GSEA_BP$NAME)
GSEA_CC$NAME <- gsub("GOCC_","",GSEA_CC$NAME)
GSEA_MF$NAME <- gsub("GOMF_","",GSEA_MF$NAME)
GSEA_REACTOME$NAME <- gsub("REACTOME_","",GSEA_REACTOME$NAME)

GSEA_HALLMARK$NAME <- gsub("_"," ",GSEA_HALLMARK$NAME)
GSEA_BP$NAME <- gsub("_"," ",GSEA_BP$NAME)
GSEA_CC$NAME <- gsub("_"," ",GSEA_CC$NAME)
GSEA_MF$NAME <- gsub("_"," ",GSEA_MF$NAME)
GSEA_REACTOME$NAME <- gsub("_"," ",GSEA_REACTOME$NAME)

#| Save tables
writexl::write_xlsx(GSEA_HALLMARK, "Hallmark.xlsx")
writexl::write_xlsx(GSEA_BP, "GO-BP.xlsx")
writexl::write_xlsx(GSEA_CC, "GO-CC.xlsx")
writexl::write_xlsx(GSEA_MF, "GO-MF.xlsx")
writexl::write_xlsx(GSEA_REACTOME, "REACTOME.xlsx")
################################################################################


################################################################################
#| HEAT MAP BY DATASETTS 
################################################################################
vector <- c("HALLMARK", "BP", "CC", "MF", "REACTOME")
n <- 2
data_PTEN_loss <- data.frame(value = c(rep("PTEN loss", times =n), rep("PTEN loss", times =n), rep("PTEN loss", times =n), rep("PTEN loss", times =n), rep("PTEN loss", times =n)),
                             Name = c(GSEA_HALLMARK$NAME[1:n],GSEA_BP$NAME[1:n], GSEA_CC$NAME[1:n], GSEA_MF$NAME[1:n],REACTOME = GSEA_REACTOME$NAME[1:n]),
                             NES = c(GSEA_HALLMARK$NES[1:n],GSEA_BP$NES[1:n], GSEA_CC$NES[1:n],GSEA_MF$NES[1:n],GSEA_REACTOME$NES[1:n]),
                             dataset = c(rep("HM", times =n), rep("BP", times =n), rep("CC", times =n), rep("MF", times =n), rep("REAC", times =n)))
name <- c()
for (i in 1:length(data_PTEN_loss$Name)){
  l <- tolower(data_PTEN_loss$Name)
  name<- c(name,paste(toupper(substring(l[i],1,1)), substring(l[i],2), sep =""))
  
}

data_PTEN_loss$Name <- name
data_PTEN_loss$dataset <- factor(data_PTEN_loss$dataset, levels = c("HM", "CC", "BP", "MF", "REAC"))

ggplot(data_PTEN_loss,aes(dataset, reorder(Name, NES), color= NES)) + 
  geom_point(size=8) +
  theme(text=element_text(size=12,  family="sans"), 
        legend.key.size = unit(0.8, 'cm'), 
        axis.text.y = element_text(size = 12, color ="black"),
        axis.text.x = element_text(size = 12, color ="black"),
        plot.title=element_text(size=13, hjust = 0.5))+
  scale_color_viridis(option="C" ) +
  ylab("Term name") + 
  xlab("Datasets") +
  grids(linetype = "dashed")
ggsave("HALLMARK_BP_CC_MF_REACTOME_dotplot_less_new.pdf", height = 4.5, width = 6.5)
################################################################################


################################################################################
#| HEATMAPS BY DATASETS SEPARATELY 
################################################################################

#| HALLMARK
GSEA_HALLMARK <- GSEA_HALLMARK[order(GSEA_HALLMARK$NES, decreasing =T),]
ggplot(GSEA_HALLMARK[1:10,], aes(x= NES, y= reorder(NAME, NES)) ) +
  geom_bar(stat="identity", fill =rainbow(5)[1], color ="black") +
  theme(text=element_text(size=12,  family="sans"), 
        legend.key.size = unit(0.8, 'cm'), 
        axis.text.y = element_text(size = 8, color ="black"),
        plot.title=element_text(size=13, hjust = 0.5, face ="bold"))+
  ylab("Datasets") +
  xlab("NES")
ggsave("HALLMARK_barplot.pdf", heigh=3.5, width = 5)

GSEA_HALLMARK$HALLMARK <- "HALLMARK"
ggplot(GSEA_HALLMARK[1:10,], aes(x= HALLMARK, y= reorder(NAME, NES)) ) +
  geom_point( aes( size =NES), fill = rainbow(5)[1], shape = 21) +
  theme(text=element_text(size=12,  family="sans"), 
        legend.key.size = unit(0.8, 'cm'), 
        axis.text.y = element_text(size = 8, color ="black"),
        plot.title=element_text(size=13, hjust = 0.5, face ="bold"))+
  ylab("Datasets") +
  xlab("")
ggsave("HALLMARK_bubbleplot.pdf", heigh=3.5, width = 4)


#| BP
GSEA_BP <- GSEA_BP[order(GSEA_BP$NES, decreasing =T),]
ggplot(GSEA_BP[1:10,], aes(x= NES, y= reorder(NAME, NES)) ) +
  geom_bar(stat="identity", fill =rainbow(5)[2], color ="black") +
  theme(text=element_text(size=12,  family="sans"), 
        legend.key.size = unit(0.8, 'cm'), 
        axis.text.y = element_text(size = 8, color ="black"),
        plot.title=element_text(size=13, hjust = 0.5, face ="bold"))+
  ylab("Datasets") +
  xlab("NES")
ggsave("BP_barplot.pdf", heigh=3.5, width = 5)

GSEA_BP$BP <- "GO:BP"
ggplot(GSEA_BP[1:10,], aes(x= BP, y= reorder(NAME, NES)) ) +
  geom_point( aes( size =NES), fill = rainbow(5)[2], shape = 21) +
  theme(text=element_text(size=12,  family="sans"), 
        legend.key.size = unit(0.8, 'cm'), 
        axis.text.y = element_text(size = 8, color ="black"),
        plot.title=element_text(size=13, hjust = 0.5, face ="bold"))+
  ylab("Datasets") +
  xlab("")
ggsave("GO-BP_bubbleplot.pdf", heigh=3.5, width = 4)


#| CC
GSEA_CC <- GSEA_CC[order(GSEA_CC$NES, decreasing =T),]
ggplot(GSEA_CC[1:10,], aes(x= NES, y= reorder(NAME, NES)) ) +
  geom_bar(stat="identity", fill =rainbow(5)[3], color ="black") +
  theme(text=element_text(size=12,  family="sans"), 
        legend.key.size = unit(0.8, 'cm'), 
        axis.text.y = element_text(size = 8, color ="black"),
        plot.title=element_text(size=13, hjust = 0.5, face ="bold"))+
  ylab("Datasets") +
  xlab("NES")
ggsave("CC_barplot.pdf", heigh=3.5, width = 5)

GSEA_CC$CC <- "GO:CC"
ggplot(GSEA_CC[1:10,], aes(x= CC, y= reorder(NAME, NES)) ) +
  geom_point( aes( size =NES), fill = rainbow(5)[3], shape = 21) +
  theme(text=element_text(size=12,  family="sans"), 
        legend.key.size = unit(0.8, 'cm'), 
        axis.text.y = element_text(size = 8, color ="black"),
        plot.title=element_text(size=13, hjust = 0.5, face ="bold"))+
  ylab("Datasets") +
  xlab("")
ggsave("GO-CC_bubbleplot.pdf", heigh=3.5, width = 4.5)


#| MF
GSEA_MF <- GSEA_MF[order(GSEA_MF$NES, decreasing =T),]
ggplot(GSEA_MF[1:10,], aes(x= NES, y= reorder(NAME, NES)) ) +
  geom_bar(stat="identity", fill =rainbow(5)[5], color ="black") +
  theme(text=element_text(size=12,  family="sans"), 
        legend.key.size = unit(0.8, 'cm'), 
        axis.text.y = element_text(size = 8, color ="black"),
        plot.title=element_text(size=13, hjust = 0.5, face ="bold"))+
  ylab("Datasets") +
  xlab("NES")
ggsave("MF_barplot.pdf", heigh=3.5, width = 5)

GSEA_MF$MF <- "GO:MF"
ggplot(GSEA_MF[1:10,], aes(x= MF, y= reorder(NAME, NES)) ) +
  geom_point( aes( size =NES), fill = rainbow(5)[5], shape = 21) +
  theme(text=element_text(size=12,  family="sans"), 
        legend.key.size = unit(0.8, 'cm'), 
        axis.text.y = element_text(size = 8, color ="black"),
        plot.title=element_text(size=13, hjust = 0.5, face ="bold"))+
  ylab("Datasets") +
  xlab("")
ggsave("GO-MF_bubbleplot.pdf", heigh=3.5, width = 4.5)


#| REACTOME
GSEA_REACTOME <- GSEA_REACTOME[order(GSEA_REACTOME$NES, decreasing =T),]
ggplot(GSEA_REACTOME[1:10,], aes(x= NES, y= reorder(NAME, NES)) ) +
  geom_bar(stat="identity", fill =rainbow(5)[4], color ="black") +
  theme(text=element_text(size=12,  family="sans"), 
        legend.key.size = unit(0.8, 'cm'), 
        axis.text.y = element_text(size = 8, color ="black"),
        plot.title=element_text(size=13, hjust = 0.5, face ="bold"))+
  ylab("Datasets") +
  xlab("NES")
ggsave("REACTOME_barplot.pdf", heigh=3.5, width = 5)

GSEA_REACTOME$REAC <- "REACTOME"
ggplot(GSEA_REACTOME[1:10,], aes(x= REAC, y= reorder(NAME, NES)) ) +
  geom_point( aes( size =NES), fill = rainbow(5)[4], shape = 21) +
  theme(text=element_text(size=12,  family="sans"), 
        legend.key.size = unit(0.8, 'cm'), 
        axis.text.y = element_text(size = 8, color ="black"),
        plot.title=element_text(size=13, hjust = 0.5, face ="bold"))+
  ylab("Datasets") +
  xlab("")
ggsave("REACTOME_bubbleplot.pdf", heigh=3.5, width = 4.5)

writexl::write_xlsx(GSEA_REACTOME, "Supplementary_Table2.xlsx")
################################################################################
